Check binary image existence when ready to push
KONFLUX-8161

When ready to push the built source image, make sure the binary image is
still valid. This behavior addresses a scenarior using floating tag in
Component image repositories. Here is an example of output-image passed
to build pipeline:

  - name: output-image
    value: quay.io/org/registry-operator:{{target_branch}}

this results in potential race condition when two PipelineRuns run in
parallel and the latest source container image can be overwritten by the
older one built by the other PipelineRun.

To solve this issue, this solution introduces the following major changes:

* --output-binary-image is renamed to --binary-image-ref which accepts
  an image reference with at least digest.
* A new class ImageRef helping validating and constructing kinds of
  image reference on-demand.
* Check input binary image just before pushing the source image
* Build result has a new status named "drop". Then, when the input
  binary image does not exist anymore in the registry, stop the process
  immediately without push and build result has status "drop" and a
  meaningful message.